Update some DrsObjectForm logic and use a hook to update active DRS Object when navigating to NewDrs page
