+++ src/components/drs/pages/NewDrs.js
@@ -1,5 +1,5 @@
 import '@fontsource/roboto';
-import React from 'react';
+import React, {useEffect} from 'react';
 import { 
     Typography, 
     Container, 
@@ -10,12 +10,56 @@ import DrsObjectForm from '../DrsObjectForm';
 import {
     Link
 } from "react-router-dom";
+import { format } from 'date-fns';
 
 /* Render NewDrs page */
 const NewDrs = (props) => {
     console.log('new drs page');
     let activeDrsObject = props.activeDrsObject;
     console.log(activeDrsObject);
+
+    useEffect(() => {
+        console.log('new drs object effect');
+        let newDate = new Date();
+        newDate.setSeconds(0, 0);
+        let year = newDate.getUTCFullYear();
+        let month = newDate.getUTCMonth();
+        let date = newDate.getUTCDate();
+        let hours = newDate.getUTCHours();
+        let minutes = newDate.getUTCMinutes();
+        let seconds = newDate.getUTCSeconds();
+        let newDrsObject = {
+            id: '',
+            description: '',
+            created_time: format(new Date(year, month, date, hours, minutes, seconds), "yyyy-MM-dd'T'HH:mm:ss'Z'"),
+            mime_type: '',
+            name: '',
+            size: '',
+            updated_time: format(new Date(year, month, date, hours, minutes, seconds), "yyyy-MM-dd'T'HH:mm:ss'Z'"),
+            version: '',
+            aliases: [],
+            checksums: [],
+            drs_object_children: [],
+            drs_object_parents: [],
+            file_access_objects: [],
+            aws_s3_access_objects: [],
+            is_bundle: false,
+            checksumTypes: {
+            md5: {
+                disabled: false
+            },
+            sha1: {
+                disabled: false
+            },
+            sha256: {
+                disabled: false
+            }
+            },
+            validId: false,
+            validRelatedDrsObjects: true
+        }
+        props.drsObjectFunctions.setActiveDrsObject(newDrsObject);
+    }, [])
 
     return (
         <div>
